backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
pdfplumber==0.11.4
python-docx==1.1.2
sentence-transformers==3.3.1
chromadb==0.6.3
python-dotenv==1.0.1
httpx==0.28.1
aiofiles==24.1.0
